fix(common): ParseUUIDPipe - throw exceptions with exceptionFactory only
