+++ packages/common/pipes/parse-uuid.pipe.ts
@@ -9,7 +9,13 @@ import {
   ErrorHttpStatusCode,
   HttpErrorByCode,
 } from '../utils/http-error-by-code.util';
-import { isUUID } from '../utils/is-uuid';
+
+const uuid = {
+  3: /^[0-9A-F]{8}-[0-9A-F]{4}-3[0-9A-F]{3}-[0-9A-F]{4}-[0-9A-F]{12}$/i,
+  4: /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i,
+  5: /^[0-9A-F]{8}-[0-9A-F]{4}-5[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i,
+  all: /^[0-9A-F]{8}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{12}$/i,
+};
 
 export interface ParseUUIDPipeOptions {
   version?: '3' | '4' | '5';
@@ -35,8 +41,9 @@ export class ParseUUIDPipe implements PipeTransform<string> {
       exceptionFactory ||
       (error => new HttpErrorByCode[errorHttpStatusCode](error));
   }
+
   async transform(value: string, metadata: ArgumentMetadata): Promise<string> {
-    if (!isUUID(value, this.version)) {
+    if (!this.isUUID(value, this.version)) {
       throw this.exceptionFactory(
         `Validation failed (uuid ${
           this.version ? 'v' + this.version : ''
@@ -44,5 +51,13 @@ export class ParseUUIDPipe implements PipeTransform<string> {
       );
     }
     return value;
+  }
+
+  protected isUUID(str: unknown, version = 'all') {
+    if (typeof str !== 'string') {
+      throw this.exceptionFactory('The value passed as UUID is not a string');
+    }
+    const pattern = uuid[version];
+    return pattern && pattern.test(str);
   }
 }
