--- packages/common/utils/is-uuid.ts
@@ -1,17 +1,0 @@
-import { BadRequestException } from '../exceptions';
-import { isString } from './shared.utils';
-
-const uuid = {
-  3: /^[0-9A-F]{8}-[0-9A-F]{4}-3[0-9A-F]{3}-[0-9A-F]{4}-[0-9A-F]{12}$/i,
-  4: /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i,
-  5: /^[0-9A-F]{8}-[0-9A-F]{4}-5[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i,
-  all: /^[0-9A-F]{8}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{12}$/i,
-};
-
-export function isUUID(str: any, version = 'all') {
-  if (!isString(str)) {
-    throw new BadRequestException('The value passed as UUID is not a string');
-  }
-  const pattern = uuid[version];
-  return pattern && pattern.test(str);
-}
